pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "vidya"
version = '0.1.0'
description = "Simple Motion Detection via RTSP and OpenCV"
authors = ["bobrock <bobrock@tuta.io>"]
license = "MIT"
repository = 'https://github.com/barretobrock/vidya.git'
packages = [
    { include = 'vidya' },
]
include = ["CHANGELOG.md"]

[tool.poetry.dependencies]
python = ">=3.12.4,<4.0"
Flask = "^3"
gunicorn = "^23"
imageio = {extras = ["ffmpeg"], version = "^2.37"}
imutils = "^0.5"
loguru = "^0"
numpy = "^2.2"
opencv-contrib-python = "^4.11"
Pillow = "^11.2"
pygifsicle = "^1.1"
python-dotenv = "^1"
requests = "^2.32"
slack_sdk = "^3.35"

# Optional dependencies would go down here
# example = { version = ">=1.7.0", optional = true }

[tool.poetry.group.dev.dependencies]
pre-commit = "^3"
pytest = "^8"
pytest-cov = "^6"
flake8 = "^7"
tox = "^4"

[tool.poetry.extras]
test = ["pytest"]

[tool.isort]
profile = 'black'
force_grid_wrap = 2
force_sort_within_sections = true
skip_gitignore = true
skip_glob = ['*_pb2.py', '*_pb2_grpc.py']
